--- Happenings-NUS/happenings-NUS/src/main/java/happenings_NUS/happenings_NUS/App.java
@@ -1,13 +1,0 @@
-package happenings_NUS.happenings_NUS;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
