--- README.md
@@ -1,2 +1,0 @@
-# Happenings-NUS
-A bot to filter out emails of NUS students
