+++ happenings-NUS/src/main/java/main/happenings_NUS/App.java
@@ -1,0 +1,13 @@
+package main.happenings_NUS;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
